backend/requirements.txt
fastapi
uvicorn
pandas
biopython
python-multipart
httpx
